+++ TaskInstance.h
@@ -14,7 +14,7 @@ using namespace std;
 #define CV1_BOARD_H
 
 
-static const short ARR_INIT_SIZE = 10;
+static const int ARR_INIT_SIZE = 10;
 
 static const char EMPTY_SQUARE = '0';
 static const char BLACK_PEON = '1';
@@ -27,20 +27,20 @@ public:
     TaskInstance();
     TaskInstance(const TaskInstance &old);
 
-    short movesCount; // number of moves made by queen
-    short blacksCount; // sum of present black peons
+    int movesCount; // number of moves made by queen
+    int blacksCount; // sum of present black peons
 
-    pair<short, short> queenPosition;
+    pair<int, int> queenPosition;
     char board[ARR_INIT_SIZE][ARR_INIT_SIZE];
 
-    void readFromFile(const string &filename, short &k, short &h);
-    void getPossibleMoves(short const &k, vector<pair<short, short>> &possibleMoves);
+    void readFromFile(const string &filename, int &k, int &h);
+    void getPossibleMoves(int const &k, vector<pair<int, int>> &possibleMoves);
 
-    void printTaskInfo(short &k, short &h);
-    void printBoard(short &k);
+    void printTaskInfo(int &k, int &h);
+    void printBoard(int &k);
 
 private:
-    bool shouldMoveFurther(short const &x, short const &y, vector<pair<short, short>> &possibleMoves);
+    bool shouldMoveFurther(int const &x, int const &y, vector<pair<int, int>> &possibleMoves);
 };
